Add simple bitcoin app to PoC
This app creates the ECDSA key pair for a given private key and logs
uncompressed/compressed public key and X and Y coordinates of public
key. It also signs a Bitcoin message and logs message signature and
signature parameters such as: v, r, s and recovery id. Thse parameters
can be used in the smart contract unit tests that proves that PoC works
as expected.