Get rid off linting config per package
Since we configured the linting config for the entire monorepo there is
no need to keep linting per package. We want to override some config for
`core` package so we add the `.eslintrc` file to set custom rule for
this package.

Here we temporarily switch off pre-commit config - we are going to
address it in follow-up work.