Add a script that formats code
Add a script that formats code for a given workspace based on the global
linting configuration. Available command line arguments:
- `--workspace` or `-w`: defines the workspace name eg. core for which
  the code formatting will be invoked. This argument is required,
- `--type` or `-t`: Determines the file type. If not defined all
  supported types will be formatted,
- `--fix` or `-f`: should automatically fix linting errors.

We want to run linting in pre-commit hook but only for those pacakges
where changes are taking place. Eg. when working on `website` the
pre-commit hook should only be invoked for this workspace not for the
others. This script allows us to run linting for different file types:
js/ts, sol and config files and, only for a given workspace.