Fix the prettier configuration issue
The prettier is config in the `dapp/.prettierrc.js`  file. This configuration depends on the type value in your `package.json`. Adding "type": "module" in package.json makes all `.js` files ES modules, and then they can only be imported in other ES modules. Prettier is not able to import `.prettierrc.js` because it is now an ES module.