Copying ERC4626 related contracts from ERC4626-Alliance
Only some of the contracts were copied and pasted just for Acre needs.
In Acre project we use HH with yarn manager and this is why the path to
solmate's library in node_modules had to be changed. Added only `/src` to the path.
The rest of the code was unchanged and is based on 643cd04 commit from the
following repo: https://github.com/ERC4626-Alliance/ERC4626-Contracts