Define .npmrc in the root directory
When npmrc was defined with public hoist in the `dapp` directory it
required installation of the dependencies to happen inside the `dapp`
directory to correctly resolve the `@chakra-ui/theme-tools` dependency.

With moving the file to the root directory, installation of the
dependencies inside the root directory will correctly resolve
`@chakra-ui` dependencies inside the `dapp` sub-project.