Improve the use of  `VStack`, `HStack` and `Flex`
The Stack component and the Flex component have their children spaced out evenly but the key difference is that the Stack won't span the entire width of the container whereas the Flex will.

Let's use the Stack component when it is a simple row or column of items.  If we need to customize the items let's use Flex.

More info: https://chakra-ui.com/docs/components/stack#notes-on-stack-vs-flex