Add `Staked` event
Instead of this require we want to emit an event that will include the
referral. The off-chain accounting will need it to calculate the fee
shares. To not duplicate the `Deposit` event from `ERC4626` the `Staked`
event has only context for referral program. We can get other values
from the `Deposit` event since it will be emitted in the same
transaction.