Revert "Add support for receive approval pattern to Acre"
This reverts commit f36c65cd38ff7ba6a2e04944671f4d4523c294aa.

To support receive approval pattern we need to override the
`ERC4626.deposit` function to allow pass an `owner`. The OZ
implementation sets the owner as `msg.sender` by default under the hood
see
https://github.com/OpenZeppelin/openzeppelin-contracts/blob/master/contracts/token/ERC20/extensions/ERC4626.sol#L178
and we can't change the owner. W/o this update using approve and call
pattern to stake tokens, the `msg.sender` is tBTC token address, so we
are actually trying to send tokens from the token contract to the
receiver, which is impossible. We are going to address it in a separate
PR.