Update unit test for staking
To be closer to a real world scenarios we get rid of infinite approval
dor staking contract and approve an amount that will be staked - just to
make sure Acre can't spend more than it was approved for. Here we also
add additional test case where staker tries to stake more than approved.