Update `maxMint` function
According to the EIP-4626 `previewDeposit` and `convertToShares`
slightly differ, especially in terms of taking fees into account. We
plan to add fees for deposits and withdrawals that will be included in
the previewDeposit function. This will cause `previewDeposit` and
`convertToShares` return different values. Here for the `maxMint`
calculation, we want to have pure assets to shares conversion, exclusive
of the fees.