Implement tBTC Depositor contract
The tBTC Depositor contract is an implementation of a depositor contract
mentioned in
[RFC-11](https://github.com/keep-network/tbtc-v2/blob/main/docs/rfc/rfc-11.adoc).

This contract serves as an integrator of tBTC Bridge and Acre staking
contract.

Bitcoin deposits revealed via tBTC Depositor contract will be
automatically staked in Acre after tBTC minting process is completed on
the tBTC network side.

The process consists of two steps:
- initializeStake - this step reveals a deposit to tBTC Bridge to start
  minting process
- finalizeStake - this step should be called after tBTC minting was
  completed to stake the tBTC tokens in Acre

The functions are unprotected, it means that anyone can call them (e.g.
bots). This solution will be used to enable gasless minting experience
to the users, where only funding Bitcoin transaction will be required
from them.

The complexity comes with the actual minted tBTC amount calculation, as
the fees are not unambiguously predictable.
The stake finalization step calculates the amount to stake in Acre contract by deducting
tBTC network minting fees from the initial funding transaction amount.
The amount to stake is calculated depending on the process the tBTC
was minted in:
- for swept deposits:
  `amount = depositAmount - depositTreasuryFee - depositTxMaxFee`
- for optimistically minted deposits:
  ```
  amount = depositAmount - depositTreasuryFee - depositTxMaxFee
           - optimisticMintingFee
  ```
These calculation are simplified and can leave some positive
imbalance in the Depositor contract.
- depositTxMaxFee - this is a maximum transaction fee that can be deducted
  on Bitcoin transaction sweeping,
- optimisticMintingFee - this is a optimistic minting fee snapshotted
  at the moment of the deposit reveal, there is a chance that the fee
  parameter is updated in the tBTC Vault contract before the optimistic
  minting is finalized.
The imbalance should be donated to the Acre staking contract by the
Governance.