Simplify the type logic for the amount
The amount for `CurrencyBalance` can be of string or number type. We use numbers in bigint format, which should be converted to the correct human format. However, we should allow the use of the `CurrencyBalance` component in case we don't need the formatting (simple value in USD). For this, let's use the variable `shouldBeFormatted` and display it as a local string.