Add unit tests for `deposit` function
We overridden `deposit` function and now it takes into account the
minimum deposit amount parameter - here we add unit test for this
function to cover a cases where the deposit amount is less than minimum
and equal to the minimum amount. We indirectly cover this check's test
in `stake` tests but we want to add an explicit `deposit` function test
that covers it as well.