Create `ChainMessages` interface
The main idea is to hide the signing process under the exact
implementation for a given chain. For example, we can use `ethers` lib
for Ethereum implementation, but different lib for other chains. The
Acre staking flow will require a signature and the signig message
implementation will be used in the staking module.