Define tBTC Depositor's depositor fee
Define a fee for tBTC Depositor contract that will be taken before
staking tBTC in Acre contract.
The fee should cover costs of running maintainer bots that will execute
initializeStake and finalizeStake methods.

The fee is calculated based on the divisor.
Initial value for the fee is set to zero.