Updating vaults weight
Each vault will have a weight assigned to it. TBTC assets will be
distributed from the Dispatcher contract based on the vault's weight.
E.g. if the vault's weight is 40%, then 40% of all TBTC in Acre system
will be transferred to that vault.

An owner will have to provide a vault's weight when authorizing a new
vault. An owner can also update weights on all the authorized vaults.
Vaults total weight cannot exceed 10000 -> 100%. This number is chosen for
precision purposes. If an owner decides to assign e.g. 15.25% to a given
vault, then the vault's weight should be set to 1525.