Use `jest` instead of `chai` and `mocha`
`Jest` testing framework has built-in support for mocks, stubs, and
spies. Chai requires the installation of additional dependencies but
most of them are no longer maintained. I also tried using the `chai`
with `ethereum-waffle` package, that adds support for mocking contracts,
but it only supports ethers in v5, we want to use the latest version of
ethers - it is v6.