Adding fee calculation for deposit and mint functions
When a user calls a deposit function providing assets amount, the fee
should be included in that amount. Meaning that the actual deposit will
be less, ie deposit(assets - fee, receiverOfShares)

When a user calls a mint function providing total shares that should be
transferred to his account, a total approved assets should also include
fee. A `previewMint(shares)` function can be called to find out how much
assets should be approved and transferred in order to receive `shares`
amount.