Update staking module
Add `receiver` and `referral` params to the `initializeStake` function.
We need to pass these values as parsed hex as extra data to the tBTC
system. Currently left only TODO because we need to wait for extra data
feature in tbtc-v2.ts lib. Also we decided to not include the signing
process here but the `StakeInitialization` should be responsible for
this.