Add `signMessage` step to `StakeInitialization`
User shuold sign message before the stake. We decided to implelemnt it
as a separate function that should be called before stake because we
want to give developers more flexibility. It's not a part of `stake`
function because e.g. in the dapp after signing message there may be
some additional steps. After signing we save the signed message in the
instance of `StakeInitialization` which then can be used inside `stake`
function w/o passing message to this function.