Leave `TODO` comment in `stake` fn
Currently to do not block the integration SDK <-> dApp we want to send
Ethereum transaction as part of the staking flow. In the final version
users won't have to sign the Ethereum transaction, they will just send
stake details to a relayer bot which will stake BTC on user's behalf.