Reverting mint function to the previous state
We ageed that minimumDepositAmount should also account for treasury
fees. This decision has two sides. On the one hand a user deposits an
amount and can expect that this is what is actually staked, but in
reality the fee is subtracted first and then what's left is being
staked. On the other hand minimumDepositAmount is a fixed value (unless
changed by the owner) and won't changed if the fee is changed, which
makes it consistent value. This is how tbtcv2 works and we mimic this
behavior here.

Renamed DepositAmountLessThanMin error to LessThanMinDeposit to align
how EIP4626 name their custom errors