Let the user recall bridged tBTC
The Acre contract has a limit for the maximum number of deposits
they can accept; due to an asynchronous manner of staking where
Bitcoin has to be first bridged to tBTC via tBTC Bridge, the
limit for deposits may be reached in the meantime of stake request
initialization and finalization. In such cases, the stake request has
to wait until more deposits are accepted by the Acre contract (the
maximum limit is increased, or funds are withdrawn from the Acre
contract, making space for new deposits).
We added a path where the user is not willing to wait anymore
for the stake request to be finalized and has the possibility to
recall the tBTC stake waiting to be completed and withdraw the
liquid tBTC to their wallet.