Fix conversion of fees from sat to tBTC
depositor fee that is transferred to the treasury should be used in
`safeTransfer` in tBTC precision. So we would need to convert it twice:
```
request.amountToStake =
            (fundingTxAmount - tbtcMintingFees - depositorFee) *
            SATOSHI_MULTIPLIER;

tbtcToken.safeTransfer(acre.treasury(), depositorFee *
SATOSHI_MULTIPLIER);
```
which may be easy to confuse.
In this commit I try to conver it just once to be sure it's handled
correctly.