Create Ledger Live signer
Create a signer that sign/send transaction and sign messages using the
Ledger Live Wallet API. Thanks to this we can use the Acre SDK that is
built with ethers lib in Ledger Live dapp. Here we also create a single
instance of the Ledger Live transport that should be used across the
dapp. Otherwise if we create another instance of the transport in eg.
ledger signer class the wallet API throws an error about unhandled
request.