Drafting unstaking fees
When unstaking stBTC Acre should collect fees that goes to a fee
recipient address e.g. treasury wallet. Fees are collected duriing
redeeming (providing a concrete amount of shares to convert for tBTC)
or withdrawing (providing a concrete amount of assets to collect and
burn shares). Exit fees are calculated using exit fee basis points that
is a governbale parameter.