Renameing Acre contract to stBTC
At first Acre was meant to serve as a main contract for the Acre
ecosystem, that is why it was named as such. However, there will be a
governance token that will be named after the project name. It can
create a lot of confusion when depositing TBTC into Acre create stBTC
shares. We decided to rename this contract to stBTC and leave the Acre
name for the governance token.