Update message signing process in stake flow
Pass the `bitcoinRecoveryAddress` to the `StakeInitialization` class.
Since we can't easily determine if we should use `true` or `false` in
`BitcoinAddressConverter.publicKeyHashToAddress` to convert public key
hash to `P2PKH` or `P2WPKH` we can use what the input was w/o doing a
conversion.

Here we also get rid of:
- the bitcoin client from the `StakeInitailization` class because at
  least for now it is unnecessary,
- `referral` field - it is decoded and passed to the contract function
  under the hood.