Handling dust for maxDeposit function
Each deposit collects some fee from a user. Max deposit returns unused
tBTC balance in the Acre system, but it does not count for the entry
fees. It results in some unsuded dust left. If the dust is below a min
amount of deposit, this function will return 0, so that users won't be
confused when they actually want to deposit below the min threshold.