Fix SDK GH worklow
We need `build` and `typechain` artifacts from `core` package in `sdk`
because we import artifacts and types generated by `typechain` in `sdk`
package. To achieve this, we extract a reusable workflow that builds the
`core` package and uploads necessary artifacts. Then we use this
reusable workflow to run successfully the `sdk` jobs.

Here we also update the `jest` config - we should ignore tests inside
`dist` dir if exists.