Implement TBTCDepositorProxy
Use TBTCDepositorProxy abstract contract implemented in https://github.com/keep-network/tbtc-v2/pull/778
as a base for the depositor contract.

The abstract contract handles calculation of the approximate bridged
tBTC amount.

The tbtcAmount caveat:

The tbtcAmount computed by _calculateTbtcAmount and returned by
_finalizeDeposit may not correspond to the actual amount of TBTC
minted for the deposit. Although the treasury fee cut upon minting
is known precisely, this is not the case for the optimistic minting
fee and the Bitcoin transaction fee. To overcome that problem,
this contract just takes the current maximum values of both fees,
at the moment of deposit finalization. For the great majority of
the deposits, such an algorithm will return a tbtcAmount slightly
lesser than the actual amount of TBTC minted for the deposit.
This will cause some TBTC to be left in the contract and ensure
there is enough liquidity to finalize the deposit. However, in
some rare cases, where the actual values of those fees change
between the deposit minting and finalization, the tbtcAmount
returned by this function may be greater than the actual amount
of TBTC minted for the deposit. If this happens and the reserve
coming from previous deposits leftovers does not provide enough
liquidity, the deposit will have to wait for finalization until
the reserve is refilled by subsequent deposits or a manual top-up.
The integrator is responsible for handling such cases.