Create a custom implementation of `DepositorProxy`
Create `RelayerDepositorProxy` that implements the `DepositorProxy`
interface to reveal a deposit via Open Zeppelin Defender Relayer bot. To
do that we need to send HTTP post request with necessary data to the
relayer under the given url. For now we hardcoded the URL but it should
be fetched from env variables. We will handle this in a follow-up work.