Combine notifyBridgingCompleted and finalizeStakeRequest
To optimize gas usage we combine finalizeStakeRequest and finalizeStakeRequest
in one function to call in one transaction.
In majority of cases both actions will be executed together, only if
deposit limit is reached in stbtc, the finalizeStakeRequest should be
called right after bridging completed and later when deposit can be
finalized in stbtc contract the finalizeStakeRequest should be called.