Fix no `ongoingRequest` error
There was an error when doing a reveal that displayed `no ongoingReques: <request_id>`. It was happening because we were using a separate transport object in our LedgerLiveEthereumSigner, and a separate one for our react hooks. The one in our react hook was always connected and we didn't disconnect it after doing a transaction. As a workaround we will save the transport object that we use in react in a separate context and we connect it only before doing a transaction, and disconnect it right after.