Introduce function to track max stake amount
The function should be used by dApp to validate maximum possible amount
of the stake request.
It takes into account limit set in the stBTC contract, maximum single
stake limit and balance of pending stakes.

The `initializeStakeRequest` function is not enforcing this
limit, not to block the reveal deposit operation of the concurrent
deposits made in the dApp in the short window between limit check,
submission of Bitcoin funding transaction and stake request
initialization.
In case of an unfortunate coincidence of concurrent deposit request
reaching the limit, the Acre protocol will work towards stake request
finalization by raising the limit to let the queued stake request in.