Fix the transaction details problem
Due to the fact that the transaction details are calculated on the basis of the amount entered by the user, we should limit these values. It should be calculated when the value is correct.Otherwise, we will show too large an amount or will not be able to display such a small amount.