Introduce finalizeStakeRequest function and queue
To optimize gas usage it is expected that after tBTC is bridged only one
function is called: `finalizeStakeRequest`.
If the function fails due to stBTC deposit limit being reached the stake
request shoudl be added to a queue with `queueForStaking` function.
When stBTC is ready to accept the deposit the `stakeFromQueue` function
should be called.