Refactor approach to function execution - staking flow
We cannot use useWalletAPIClient because there is an error from the ledger side, which does not allow us to handle success and error callbacks correctly. Therefore, we need to use the hook useSignAndBroadcastTransaction and use callbacks in the effects. To avoid repeating the logic I created the hook useExecuteFunction to easily handle callbacks. So let's use it in all cases where it is possible.