Update Ethereum impl of `ChainEIP712Signer`
In SDK, the `ChainIdentifier` interface is usually used to represent a
chain-specific address and we can only get the identifier hex w/o
prefix. The implementation of `ChainEIP712Signer` should be responsible
for adding the prefix to the value of `address` type, because the prefix
depends on the chain and it may be different. Here we use the
`getAddress` function from `ethers` lib to add `0x` prefix if needed.