Wait for Bitocin funding transaction in stake fn
We want to initiate the stake when we are sure the BTC funding
transaction is visible by Bitcoin node. The default implementation of
minting in tbtc-v2 throws an error when the deposit tx is not available
yet. In our SDK, we make sure the deposit transaction is available and
then initiate minting/staking. Here we add the back off mechanism - the
SDK retires the function that looks for a deposit transaction a
`retries` number of times. The result will return the function's return
value if no exceptions are thrown. In our case we want to try to find
the deposit transacion 4 times with a 3 seconds backoff step that will
be increased exponentially for subsequent retry attempts. If it fails
after 4 tries it will throw an error.