Store queued amount as uint88
Before StakeRequest was `uint8` + `uint160` + `uint256` which gives 2 storage slots.
However, we don't need `uint256` for `queuedAmount` as the maximum BTC supply is 21M.
That gives `21000000000000000000000000` in 1e18 precision. In practice, we can fit
everything in one slot: `uint8 + uint160 + uint88`. The `uint88` type gives us the max
value of `2^88-1 = 309485009821345068724781055`. This is more than enough to store the
entire BTC supply because
`21000000000000000000000000 < 309485009821345068724781055 = 21M < ~309M`.