Add job deploying contracts on Sepolia
We're adding a GH Actions job to the 'Core' workflow. The job will be executed
only for workflows triggered manually and will deploy `core/` contracts on
provided environment. At the moment only `sepolia` environment is supported.
Once contracts are deployed, the deployment artifacts will be exported as job
artifacts, so that they could be later committed to the repository (if that's
desired).