Upgrade actions uploading/downloading artifacts to the latest version
We had inconsistance in the used versions of the `actions/upload-artifact`
and `actions/download-artifact` actions, which caused the `Unable to download
artifact(s): Artifact not found for name: core-build` error (artifact uploaded
with the `v3` of the `actions/upload-artifact` action must be downloaded with
the `actions/download-artifact` in the same `v3` version). We're upgrading all
placec where we used `v3` to `v4` (which runs on a Node 20, in contrast to `v3`,
which uses deprecated Node 16).