Fix the upload step in the `core-deploy-testnet` job
The upload step of the `core-deploy-testnet` job was throwing a warning about no
files being uploaded. This was because we had a misconfiguration of the path. In
this commit/PR we're fixing the path and also adding a config that will cause a
job failure if no files are found by the upload step.