Handling type verification for taproot addresses
Ledger has support for taproot addresses is not supported as BTC recovery address in tBTC. When we pass such an address to the `addressToOutputScript` function we get an error. Taproot is not supported on the protocol level anyway so let's handle this check from the SDK side. If an error is catching, let's simply return false.