Implement Bitcoin Redeemer contract
The contract is used to integrate stBTC unstaking with bridging to
Bitcoin via tBTC Bridge.

The `requestRedemption` function expects the owner of stBTC shares to
approve stBTC token shares to the BticoinRedeemer contract. Then when
`requestRedemption` function is called it redeems stBTC and calls
TBTC token contract to redeem tBTC to Bitcoin.