Restructure the styles for the Alert component
According to styleguide, we need to update the styles for the Alert component. Additionally, the Toast component is based on the basic styles of Alert.

Therefore, now the Alert component is based on the styleguide. However, a separate CardAlert component was created to contain the styles of the previous alert in one place.