Create `AbstractPausable` pausable contract
This abstract contract extracts a common part of the emergency stop
mechanism. The emergency stop mechanism can be triggered by an
authorized account. Only owner of the contract can update the emergency
stop account. The child contract must override the `_checkOwner`
internal function that checks if the caller is an owner of contract and
throws an error if the sender is not the owner.