Update `estimateStakingFees` in staking module
Sum up all network fees and add total field -  we decided to add a total
field because the SDK should be responsible for summing up all fees. If
we add a new fee in the future the consumers will have to update their
code as well which is not a developer-friendly approach.