Refactor access fns in `AbstractPausable` contract
Remove the `isOwner` function and the `_onlyOwner` modifier as they were
only used in one place. There may also be a name collision with the
`Ownable` contract from the OZ library because it also defines the
`onlyOwner` modifier, so it may be confusing when and which modifier
should be used in the child contract.