Handle error occurring after the deposit action
In the event of an error occurring after the deposit transaction but before displaying the Success screen to the user, there is a low possibility of system issues. To address this, we will show an error screen prompting the user to retry. If the error is episodal, the system will recover, and upon pressing the "Retry" button, the user will see the Success screen, continuing the flow as usual. We intentionally omit a close button, guiding the user to proceed within the flow and preventing abandonment. 