Adding more tests and deployment scripts
Dispatcher in the current form will be replaced in the future. For now,
the role of the dispatcher contract will server the MezoAllocator
contract since we do not have any other L2s/DeFi partners yet. Once we
have it, it will be updated for the proper dispatcher that will include
MezoAllocator.