Adding withdrawal functionality to MezoAllocator contract
Withdrawal functionality will be based on the deposit ids. From a user
perspective it doesn't matter what id was assigned to a particular
deposit. Acre aggregates users deposits and allocate them to Mezo
Portal and then ID is assigned. When a user request a withdrawal of some
amount, Acre will try to pull funds from the latest deposit and go to
the most recent one until the requested amount is reached. All users
deposits are mixed in a bag, and rewards will be distributed and
calculated by Acre, not Mezo. LIFO is used to pull the funds from Mezo instead of
FIFO because of how the data is stored on-chain. It is simpler to pop
the latest element from the list instead of rearanging the whole deposit
ids array.