Withdrawing assets from stBTC first and then from MezoPortal deposits
Withdrawals should be satisfied first by the internal "buffer" that is
growing between the deposit allocations. If the amount to withdraw is
lower than the buffer, then only the internal accounting should be done.
However, if the buffer is lower than the withdrawal amount, then the assets
should be withdrawn from the Mezo Portal deposit(s).