Removing Dispatcher contract
Removed Dispatcher contract, Router and references from other contracts.

At the beginning Dispatcher contract will be replaced by the
MezoAllocator contract and there's no need for Dispatcher for now. Once
we have more partnerships, then we can updated Dispatcher address in
stBTC and replace it with a proper address. MezoAllocator will be one of
the supported partners for Acre.