Allocate tBTC to MezoPortal as a "rolling" deposit
Here we change the deposit approach from tracking all the Mezo Deposits
to having only a single deposit that accumulates all deposited tBTC by
Acre into a single deposit. An "old" deposit is abondoned and balance is
zeroed. A "new" deposit is created with a new id and a new balance is
equal to the old balance + new requested amount.