Allocating 100% of tBTC
Changed the current approach of allocating 100% from deposited tBTC in
stBTC. As this is MVP, a bot won't need to set an exact amount that
should be allocated in Mezo Portal.

Other changes include inlining a "free" function for some gas
optimization, removing DepositInfo struct as tracking of depositId is
enough and some other minor refactoring.