Simplify the structure of modals
After separating the `content` directory for modals, there was a problem with cycle dependencies. Therefore, all files were moved to a higher directory. This solves the problem with dependencies. In addition, it simplifies the directory structure. However, if we find a new approach in the near future we can easily update it.