Fixing withdrawal functionality in stBTC
Upon withdrawal we need to take into account fees that will be collected
by the treasury. A user will get the desired assets, but on top we need
to pull the funds from Mezo covering not only what a user wants but also
pull a bit more to cover the withdrawal fees.

Added tests to cover scenarios with allocated funds to Mezo Portal and
then withdrawing and redeeming from Acre.