Use updated TBTC module in staking process
The tBTC module wraps tBTC API and tBTC SDK and the staking process
should interact with this module.
In this commit we update references to match the new deposit
initialization in tBTC with tBTC API instead of the custom
DepositorProxy implementation.

Wait for for funding transaction helper function was moved to the tBTC
module already so here we remove it from the staking flow.