Remove Ethereum account from the wallet context
We do not need to connect Ethereum account with the dapp since we want
to operate only on the Bitcoin wallet. The Acre SDK will generate the
Ethereum address based on the Bitcoin address under the hood and the
user's identifier should always be bitcoin address and the SDK will take
a care of converting it to Ethereum address under the hood.

Still, some features in the Acre SDK expect the Ethereum address eg. to
get the estimated bitcoin balance. Hence as a temporary solution we set
the ethereum account to zero address to not break the dapp. In the
future we want to completely get rid of the ethereum account from the
dapp wallet context and we will only pass the bitcoin address to the
Acre SDK.