Update `handleDepositFinalized` handler
Throw an error when the Deposit entity deosn't exist. We throw an error
just to improve readability of the code. This should never happen
because the graph indexes events chronologically and the deposit can
only be finalized once it has been initialized. This case is possible
when we set the wrong `startBlock` in manifest but we always set the
`startBlock` to block where a given contract was deployed. Cover this
case in unit tests.