Removing whenNotPaused modifier in favor of the paused() check
The following functions: deposit, withdraw, redeem, mint already check
if the contract is paused or not. This is validated in max*
related functions downstream, e.g. maxDeposit(). If the protocol is paused, these
max* functions will return zero and stop the flow by reverting a
transaction. One of the revert errors might be e.g.
ERC4626ExceededMaxDeposit. This way we simplify the flow a little bit
and save some gas.