Add `@swan-bitcoin/xpub-lib`
To get the bitcoin address for `x'/0'/0'/0/0` path of a given extended
public key to use always the same bitocin address as an identifier
because Leder Wallet API returns the "next" public address where a user
should receive funds. In the context of Bitcoin, the address is
"renewed" each time funds are received. We need to use the same bitcoin
address to generate always the same deposit owner address on ethereum.