Use `BitcoinProvider` to initialize the Acre SDK
We want to relay on the `BitcoinProvider` implementation to get the
bitcoin address because different wallets use different strategies. For
example in Ledger Live the address is "renewed" each time funds are
received in order to allow some privacy. In that case always getting the
same bitcoin address is hidden under a given implementation.