Quick fix for transaction window pop-up
When we connected the account from the landing page the etherum account was not set up. Therefore, the deposit transaction window did not run because the eth account is still required in several places in the code.This is a quick fix to solve the issue but it is not a solution to the problem.