Fix issue with the incorrect  order of hooks for `ActionFormModal`
We were getting the error “React detected a change in the order of hooks” when we closed and opened the deposit and withdrawal flow alternately.  The problem was that we were calling `renderComponent` directly, even though it is a react component. That causes react to treat the hook calls inside the function as part of `ActionFormModal`, even though we meant for them to be part of `StakeFormModal`/`UnstakeFormModal`.

More info here: https://stackoverflow.com/questions/57397395/react-has-detected-a-change-in-the-order-of-hooks-but-hooks-seem-to-be-invoked