Remove the unneeded component
After integration with OrangeKit `UnsupportedBitcoinAddressModal` is not needed in the stake/unstake flow. The account will not be connected when the address type is not supported. Therefore, the use of this modal should be done differently. Let's remove its use for now and in the next steps we should define how we want to handle it.