Convert Bitcoin amount to stBTC and tBTC amount
The account module exposes function that initializes the withdrawal - it
expects Bitcoin amount in satoshi precision. In this case we need to
convert satoshi to stBTC shares. Also we need to estimate the amount of
tBTC released after redeeming stBTC shares including all Acre fees using
the `previewRedeem` stBTC contract function. The value is required by
the tBTC SDK to determine the tBTC Wallet from which the Bitcoin will be
bridged to the depositor's Bitcoin address.