Prototype stBTC minting by external Mezo contract
Here we prototype the idea of minting stBTC from Mezo without depositing
BTC to the vault.
In such case we need to restrict withdrawals so stBTC holders that minted
from Mezo can't withdraw BTC from the vault directly as this could lead
to draining the vault. Mezo users may be interested in withdrawing tBTC
directly from the vault, as they won't need to keep to the locktime
they have in Mezo.