Update the Withdrawal flow
Display more information about a given process to the user. To
initialize withdrawal we need to find a wallet that can hanlde the tBTC
redemption request - it's time-consuming so we display to the user that
we are building the transacion data.

The next phase is sign message step - we display a message saying please
finalize the signing process in yout wallet.

The last step is waiting for the transacion so we are saying that
withdrawal initialiazation is in progress.