Fix the issue of calling the required actions for logging in
The signing action should happen in the callback when the wallet connects correctly. This avoids races in case the user was already logged in with another wallet and quickly switches to another connector. When we call the message signing action in onSuccess we are sure that it will call for the connected connector.