Calculate withdrawal fee
We want to calculate withdrawal fee for the given satoshi amount.
We sum up the fees on tBTC and Acre protocols.
For tBTC there is a treasury fee taken on redemption,
for Acre there is an exit fee taken for the withdrawal.

We expose the fees in a similar structure we did for deposit.