Implement ERC-4626 Non-Fungible Withdrawals
This PR implements the ERC-4626 Non-Fungible Withdrawals feature. This
feature allows users to redeem their stBTC tokens in a non-fungible way,
meaning the stBTC token holder is allowed to redeem the shares only up
the the amount of the shares that were minted with a deposit operation.

If stBTC tokens holder received the tokens in a different way than by actual
depositing, they won't be able to withdraw until the non-fungible withdrawals
are disabled.

The motivation behind this feature is to mitigate the risk of the vault
draining attacks, by the users of the Mezo Portal, that would have the
possibility to mint stBTC as a receipt for their Mezo Portal deposits.

References:
- [Product Pitch](https://coda.io/d/Mezo_d_vIDNPK008/2-Drafts-Ready-for-Review_su1QZ#Draft-Product-Pitches_tuH_B/r16)
- [Prototype #485](https://github.com/thesis/acre/pull/485)